add dry-run and load_incluster_config

+++ cleanup.py
@@ -1,11 +1,18 @@
 #!/usr/bin/env python3
+import sys
+
 import urllib3
 from kubernetes import client, config
 from kubernetes.client.rest import ApiException
+import argparse
 
-urllib3.disable_warnings()
+parser = argparse.ArgumentParser(description="Pod Cleaner")
+parser.add_argument('--dry-run', action='store_true')
+args = parser.parse_args()
 
-config.load_kube_config()
+print("dry-run" if args.dry_run else 'wet-run')
+
+config.load_incluster_config()
 api = client.CoreV1Api()
 pods = api.list_pod_for_all_namespaces()
 
@@ -18,5 +25,11 @@ for pod in pods.items:
                     and container_status.last_state.terminated \
                     and container_status.last_state.terminated.reason \
                     and container_status.last_state.terminated.reason == 'ContainerCannotRun':
-                print('deleting', pod.metadata.name, 'in', pod.metadata.namespace)
-                api.delete_namespaced_pod(pod.metadata.name, pod.metadata.namespace)
+                if args.dry_run:
+                    print('dry-run: would have deleted', pod.metadata.name, 'in', pod.metadata.namespace)
+                else:
+                    try:
+                        print('deleting', pod.metadata.name, 'in', pod.metadata.namespace)
+                        api.delete_namespaced_pod(pod.metadata.name, pod.metadata.namespace)
+                    except ApiException as e:
+                        print('exception while deleting: ', e)
